Break up the GPU module into several submodules
The GPU module was getting far too large to be maintainable as a single
file, so this change breaks it up into several smaller units.
